# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: XJU Lab API
  version: 1.0.0
  description: Session-authenticated API for the XJU Lab collaboration platform.
servers:
  - url: /api/v1
...
paths:
...
  /meetings:
    get:
      tags: [meetings]
      responses: { '200': { description: Meetings visible to current member, content: { application/json: { schema: { $ref: '#/components/schemas/MeetingPage' } } } } }
    post:
      tags: [meetings]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [title, startsAt], properties: { title: { type: string }, projectId: { type: string, format: uuid, nullable: true }, startsAt: { type: string, format: date-time }, participants: { type: array, items: { type: string, format: uuid } } } } } }
      responses: { '201': { description: Created meeting, content: { application/json: { schema: { $ref: '#/components/schemas/Meeting' } } } } }
  /meetings/{meetingId}/minutes:
    get:
      tags: [meetings]
      parameters: [{ name: meetingId, in: path, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Latest minutes revision } }
    put:
      tags: [meetings]
      parameters: [{ name: meetingId, in: path, required: true, schema: { type: string, format: uuid } }, { $ref: '#/components/parameters/IfMatchVersion' }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [body], properties: { body: { type: string, maxLength: 20000 } } } } } }
      responses: { '200': { description: Saved minutes revision }, '409': { $ref: '#/components/responses/ApiError' } }
  /meetings/{meetingId}/actions:
    post:
      tags: [meetings]
      parameters: [{ name: meetingId, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [title], properties: { title: { type: string }, description: { type: string }, assigneeId: { type: string, format: uuid }, dueDate: { type: string, format: date } } } } } }
      responses: { '201': { description: Action item created as the shared task record, content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } } } }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: SESSION
    printerAgentToken:
      type: http
      scheme: bearer
  parameters:
    IfMatchVersion:
      name: If-Match-Version
      in: header
      required: true
      schema: { type: integer, minimum: 1 }
  schemas:
...
    Task:
      type: object
      required: [id, title, status, version]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid, nullable: true }
        title: { type: string }
        description: { type: string }
        assigneeId: { type: string, format: uuid, nullable: true }
        createdBy: { type: string, format: uuid }
        dueDate: { type: string, format: date, nullable: true }
        status: { type: string, enum: [OPEN, IN_PROGRESS, DONE, CANCELED] }
        version: { type: integer }
    Meeting:
      type: object
      required: [id, title, startsAt, createdBy, version]
      properties:
        id: { type: string, format: uuid }
        projectId: { type: string, format: uuid, nullable: true }
        title: { type: string }
        startsAt: { type: string, format: date-time }
        createdBy: { type: string, format: uuid }
        version: { type: integer }
...
    MeetingPage: { $ref: '#/components/schemas/PageEnvelopeMeeting' }
...
    PageEnvelopeBase:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { type: object } }
        total: { type: integer, format: int64 }
        page: { type: integer }
        pageSize: { type: integer }
...
    PageEnvelopeMeeting:
      allOf:
        - $ref: '#/components/schemas/PageEnvelopeBase'
        - type: object
          properties: { items: { type: array, items: { $ref: '#/components/schemas/Meeting' } } }
...
  responses:
    ApiError:
      description: Consistent request error
      content:
        application/json:
          schema:
            type: object
            required: [code, message, requestId]
            properties:
              code: { type: string }
              message: { type: string }
              requestId: { type: string }
              fieldErrors: { type: object, additionalProperties: { type: string } }
security:
  - sessionCookie: []